// Frontend/my-app/src/common/define-type.ts
export interface GarbageQuantity{
    KhoiLuong: number,
    NhanRac: string
}
export interface BarChartData{
    Ngay: string;
    Rac: GarbageQuantity[]
}